ログイン直後に、ログインしたユーザー名が表示されるようにした
