+++ NyanNyanEngine/model/repository/TweetsRepository.swift
@@ -13,7 +13,6 @@ import RxRelay
 protocol BaseTweetsRepository: AnyObject {
     func isLoggedIn() -> Bool
     func getHomeTimeLine(uiRefreshControl: UIRefreshControl?) -> Observable<[Status]?>
-    func getCurrentUser() -> Observable<String>
     
     var statuses: Observable<[Status]?> { get }
     var buttonRefreshExecutedAt: AnyObserver<String>? { get }
@@ -73,14 +72,6 @@ class TweetsRepository: BaseTweetsRepository {
         return self.apiClient
             .postResponse(urlRequest: urlRequest)
             .map { [unowned self] in self.toStatuses(data: $0) }
-    }
-    
-    func getCurrentUser() -> Observable<String> {
-        let tekitou = userDefaultsConnector.getString(withKey: "screen_name") ?? "にゃんにゃんエンジン"
-        return Observable<String>.create { observer in
-            observer.onNext(tekitou)
-            return Disposables.create()
-        }
     }
     
     func isLoggedIn() -> Bool {
